Implementado rota e model de feitiços

+++ api/models/Character.js
@@ -1,5 +1,6 @@
-const DB = require('../helpers/Db')
+const DB = require('../helpers/Db.js')
 const Item = require('./Item.js')
+const Spell = require('./Spell.js')
 
 exports.read = (next) => {
     var script = "SELECT * FROM tbPersonages"
@@ -27,7 +28,18 @@ exports.read = (next) => {
                 dexterity: results[r].dexterity,
                 inteligence: results[r].inteligence,
                 wisdom: results[r].wisdom,
-                charisma: results[r].charisma
+                charisma: results[r].charisma,
+                spellSlots: [
+                    results[r].spellSlot1,
+                    results[r].spellSlot2,
+                    results[r].spellSlot3,
+                    results[r].spellSlot4,
+                    results[r].spellSlot5,
+                    results[r].spellSlot6,
+                    results[r].spellSlot7,
+                    results[r].spellSlot8,
+                    results[r].spellSlot9,
+                ]
             }
             console.log(character)
             chars.push(character)
@@ -38,11 +50,15 @@ exports.read = (next) => {
 
 exports.create = (char, next) => {
     var script = `INSERT INTO tbPersonages (charname, gender, alignment, age, race, class, `+
-    `lvl, xp, hitDice, proficiency, strenght, constitution, dexterity, inteligence, wisdom, charisma) `+
+    `lvl, xp, hitDice, proficiency, strenght, constitution, dexterity, inteligence, wisdom, charisma, `+
+    `spellSLots1, spellSLots2, spellSLots3, spellSLots4, spellSLots5, `+
+    `spellSLots6, spellSLots7, spellSLots8, spellSLots9)`+
     `VALUES ("${char.name}", ${char.gender}, ${char.alignment}, ${char.age}, "${char.race}", `+
     `"${char.class}", ${char.level}, ${char.xp}, ${char.hitDice}, ${char.proficiency}, `+
     `${char.strenght}, ${char.constitution}, ${char.dexterity}, ${char.inteligence}, `+
-    `${char.wisdom}, ${char.charisma})`
+    `${char.wisdom}, ${char.charisma}, ${char.spellSlots[0]}, ${char.spellSlots[1]}, `+
+    `${char.spellSlots[2]}, ${char.spellSlots[3]}, ${char.spellSlots[4]}, ${char.spellSlots[5]}, `+
+    `${char.spellSlots[6]}, ${char.spellSlots[7]}, ${char.spellSlots[8]})`
     console.log(script)
     DB.con().query(script, (err, results, fields) => {
         if(err){
@@ -62,7 +78,12 @@ exports.update = (char, next) => {
     `age = ${char.age}, race = "${char.race}", class = "${char.class}", lvl = ${char.level}, `+
     `xp = ${char.xp}, hitDice = ${char.hitDice}, proficiency = ${char.proficiency}, `+
     `strenght = ${char.strenght}, constitution = ${char.constitution}, dexterity = ${char.dexterity}, `+
-    `inteligence = ${char.inteligence}, wisdom = ${char.wisdom}, charisma = ${char.charisma} `+
+    `inteligence = ${char.inteligence}, wisdom = ${char.wisdom}, charisma = ${char.charisma}, `+
+    `spellSlot1 = ${char.spellSlots[0]}, spellSlot2 = ${char.spellSlots[1]}, `+
+    `spellSlot3 = ${char.spellSlots[2]}, spellSlot4 = ${char.spellSlots[3]}, `+
+    `spellSlot5 = ${char.spellSlots[4]}, spellSlot6 = ${char.spellSlots[5]}, `+
+    `spellSlot7 = ${char.spellSlots[6]}, spellSlot8 = ${char.spellSlots[7]}, `+
+    `spellSlot9 = ${char.spellSlots[8]} `+
     `WHERE id = ${char.id}`
     console.log(script)
     next(true)
@@ -79,18 +100,31 @@ exports.update = (char, next) => {
 }
 
 exports.delete = (charId, next) => {
-    var charItems = Item.deleteCharItems(charId, (deleted) => {     //Delete character items first
-        //Delete character:
-        var script = "DELETE FROM tbPersonages WHERE id = " + charId
-        DB.con().query(script, (err, results, fields) => {
-            if(err){
-                console.log(err)
-                next(false)
-            }
-            else{
-                console.log('personagem deletado com sucesso')
-                next(true)
-            }
-        })
+    Item.deleteCharItems(charId, (deletedItems) => {     //Delete character items first
+        if(deletedItems)
+            Spell.deleteCharSpells(charId, (deletedSpells) => {
+                //Finally, delete character:
+                if(deletedSpells){
+                    var script = "DELETE FROM tbPersonages WHERE id = " + charId
+                    DB.con().query(script, (err, results, fields) => {
+                        if(err){
+                            console.log(err)
+                            next(false)
+                        }
+                        else{
+                            console.log('personagem deletado com sucesso')
+                            next(true)
+                        }
+                    })
+                }
+                else{
+                    console.log('abortando')
+                    next(false)
+                }
+            })
+        else{
+            console.log('abortando')
+            next(false)
+        }
     })
 }
